requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
Faker==11.3.0
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
redis==4.1.0
requests==2.27.1
text-unidecode==1.3
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3